refactor: modularize footer components and update structure for improved readability
