+++ index.html
@@ -2,7 +2,11 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/png" href="src/assets/images/poke.png" />
+    <link
+      rel="icon"
+      type="image/png"
+      href="src/assets/logos/poke-mart-logo.png"
+    />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="/src/styles/main.scss" />
     <title>Poke Webshop</title>
