+++ src/components/footer/FooterLogo.tsx
@@ -1,0 +1,17 @@
+import { Heading, Text } from "../../utils";
+import styles from "../../styles/components/common/Footer.module.scss";
+
+const FooterLogo = () => (
+  <div className={styles.logoSection}>
+    <Heading size="headingxl" className={styles.footerLogo}>
+      Poke-Mart
+    </Heading>
+    <Text className={styles.address}>
+      Oude Baan 2, 2800 Mechelen
+      <br />
+      Belgium
+    </Text>
+  </div>
+);
+
+export default FooterLogo;
