+++ src/components/footer/index.ts
@@ -1,0 +1,5 @@
+export { default as Footer } from "./Footer.tsx";
+export { default as FooterLinks } from "./FooterLinks.tsx";
+export { default as Newsletter } from "./Newletter.tsx";
+export { default as Copyright } from "./Copyright.tsx";
+export { default as FooterLogo } from "./FooterLogo.tsx";
