+++ src/utils/index.ts
@@ -4,6 +4,9 @@ export * from "../components/common";
 // Header
 export * from "../components/header";
 
+// Footer
+export * from "../components/footer";
+
 // Layout
 export * from "../pages";
 
@@ -13,7 +16,7 @@ export * from "../components/cart";
 // Home
 export * from "../components/home";
 
-// filters
+// filter
 export * from "../components/filter";
 export * from "./helperFunctions.ts";
